Run iptables/sysctl on _every_ boot, not just first
`runcmd` will only run once per instance, but the iptables rules created
here need to run on every single boot (in the case of reboot for any
reason.) `bootcmd` is the cloudinit module we want to use.

But that runs too early (before write_files has dropped the config
files) so we need to use a trick to run the script the runcmd drops from
bootcmd.